config

+++ src/Nethermind/Nethermind.Consensus/Validators/BlockValidator.cs
@@ -435,7 +435,7 @@ public class BlockValidator(
     private bool ValidateBlockLevelAccessListSize(Block block, ref string? error)
     {
         BlockAccessList bal = block.BlockAccessList!;
-        long maxBalItems = block.Header.GasLimit / Eip7928Constants.ItemCost;
+        int maxBalItems = (int)(block.Header.GasLimit / Eip7928Constants.ItemCost);
 
         if (bal.ItemCount > maxBalItems)
         {
